+++ R/helper_assess.R
@@ -914,11 +914,12 @@ calculateSRMR <- function(
 
 
 
-#' Internal: Calculate effect size
+#' Internal: Calculate Cohens f^2
 #'
-#' Calculate the effect size for regression analysis \insertCite{Cohen1992}{cSEM}.
+#' Calculate the effect size for regression analysis \insertCite{Cohen1992}{cSEM}
+#' known as Cohen's f^2
 #'
-#' @usage calculateEffectSize(.object = NULL)
+#' @usage calculatef2(.object = NULL)
 #'
 #' @inheritParams csem_arguments
 #'
@@ -929,7 +930,7 @@ calculateSRMR <- function(
 #' @seealso [assess()], [csem], [cSEMResults]
 #' @keywords internal
 
-calculateEffectSize <- function(.object = NULL) {
+calculatef2 <- function(.object = NULL) {
   
   ## Get relevant quantities
   approach_nl      <- .object$Information$Arguments$.approach_nl
@@ -990,7 +991,7 @@ calculateEffectSize <- function(.object = NULL) {
   ## Remove 
   
   ## Make output a matrix
-  # Note: this is necessary for calculateEffectSize to work
+  # Note: this is necessary for calculatef2() to work
   #       when supplied to the .user_funs argument. Currently, .user_funs functions 
   #       need to return a vector or a matrix. I may change that in the future.
   ss <- s[vars_endo, , drop = FALSE]
@@ -1081,4 +1082,54 @@ calculateVIFModeB <- function(.object = NULL) {
   }
   
   return(VIF)
+}
+
+#' Helper for assess()
+#' @noRd
+printEffects <- function(.effect, .ci_colnames, .what = "Total effect") {
+  l <- max(nchar(.effect[, "Name"]), nchar(.what))
+  
+  if(length(.ci_colnames) != 0) {
+    xx <- regmatches(.ci_colnames, regexpr("\\.", .ci_colnames), invert = TRUE)
+    interval_names    <- unique(sapply(xx, `[`, 1))
+    sig_level_names   <- unique(gsub("[LU]", "", sapply(xx, `[`, 2)))
+    
+    cat2("\n  ",  col_align("", width = max(l, nchar(.what)) + 44))
+    for(i in interval_names) {
+      cat2(col_align(i, width = 20*length(sig_level_names), align = "center"))
+    }
+  }
+  cat2(
+    "\n  ", 
+    col_align(.what, l + 2), 
+    col_align("Estimate", 10, align = "right"), 
+    col_align("Std. error", 12, align = "right"),
+    col_align("t-stat.", 10, align = "right"), 
+    col_align("p-value", 10, align = "right")
+  )
+  if(length(.ci_colnames) != 0) {
+    for(i in rep(sig_level_names, length(interval_names))) {
+      cat2(col_align(i, 20, align = "center"))
+    } 
+  }
+  
+  for(i in 1:nrow(.effect)) {
+    cat2(
+      "\n  ", 
+      col_align(.effect[i, "Name"], l + 2), 
+      col_align(sprintf("%.4f", .effect[i, "Estimate"]), 10, align = "right"),
+      col_align(sprintf("%.4f", .effect[i, "Std_err"]), 12, align = "right"),
+      col_align(sprintf("%.4f", .effect[i, "t_stat"]), 10, align = "right"),
+      col_align(sprintf("%.4f", .effect[i, "p_value"]), 10, align = "right")
+    )
+    if(length(.ci_colnames) != 0) {
+      for(j in seq(1, length(.ci_colnames), by = 2) + 6) {
+        cat2(
+          col_align(
+            paste0("[", sprintf("%7.4f", .effect[i, j]), ";", 
+                   sprintf("%7.4f", .effect[i, j+1]), " ]"), 20, align = "center")
+        )
+      } 
+    }
+  }
 }