+++ R/postestimate_test_MICOM.R
@@ -109,8 +109,9 @@ testMICOM <- function(
     
     ### Preparation ==============================================================
     ## Get pooled data (potentially unstandardized) and remove id column
-    X <- as.matrix(.object[[1]]$Information$Data_pooled)
-    # X <- as.matrix(X[, -which(colnames(X) == .object[[1]]$Information$Arguments$.id)])
+    X <- .object[[1]]$Information$Data_pooled
+    X <- X[, -which(colnames(X) == .object[[1]]$Information$Arguments$.id)]
+    X <- as.matrix(X)
     
     # Collect initial arguments (from the first object, but could be any other)
     arguments <- .object[[1]]$Information$Arguments