+++ R/print.cSEMAssess.R
@@ -194,21 +194,21 @@ print.cSEMAssess <- function(x, ...) {
     }
   }
   
-  if(any(names(x) == "Effect_size")) {
+  if(any(names(x) == "F2")) {
     cat2("\n\n", rule2("Effect sizes (Cohen's f^2)"))
-    for(i in rownames(x$Effect_size)) {
-      ll <- nchar(colnames(x$Effect_size[i, x$Effect_size[i, ] != 0, drop = FALSE]))
+    for(i in rownames(x$F2)) {
+      ll <- nchar(colnames(x$F2[i, x$F2[i, ] != 0, drop = FALSE]))
       cat2("\n\n  Dependent construct: '", i, "'\n")
       cat2(
         "\n\t", 
         col_align("Independent construct", max(ll, nchar("Independent construct")) + 2), 
-        col_align("Effect size", 12, align = "center")
+        col_align("f^2", 12, align = "center")
       )
-      for(j in colnames(x$Effect_size[i, x$Effect_size[i, ] != 0, drop = FALSE])) {
+      for(j in colnames(x$F2[i, x$F2[i, ] != 0, drop = FALSE])) {
         cat2(
           "\n\t", 
           col_align(j, max(ll, nchar("Independent construct")) + 2), 
-          col_align(sprintf("%.4f", x$Effect_size[i, j]), 12, align = "center")
+          col_align(sprintf("%.4f", x$F2[i, j]), 12, align = "center")
         )  
       }
     }
@@ -225,22 +225,37 @@ print.cSEMAssess <- function(x, ...) {
       cat2("\n\n\tFornell-Larcker matrix\n\n")
       print(x$`Fornell-Larcker`)
     }
+  }
+  
+  if(any(names(x) == "Effects")) {
+    ### Effects ----------------------------------------------------------------
+    cat2("\n\n", rule2("Effects"), "\n\n")
     
-    # if(any(names(x) == "RA") && !anyNA(x$RA)) {
-    #   cat2("\n\n\tRedundancy analysis")
-    #   cat2(
-    #     "\n\n\t", 
-    #     col_align("Construct", max(l, nchar("Construct")) + 2),
-    #     col_align("Value", 14, align = "center")
-    #   )
-    #   for(i in names(x$RA)) {
-    #     cat2(
-    #       "\n\t", 
-    #       col_align(i, max(l, nchar("Construct")) + 2),
-    #       col_align(sprintf("%.4f",x$RA[i]), 14, align = "center")
-    #     ) 
-    #   }
-    # }
+    ## Confidence intervals
+    # Get the column names of the columns containing confidence intervals
+    ci_colnames <- colnames(x$Effects$Total_effects)[-c(1:6)]
+    
+    # Are there more confidence intervals than the default (the 95% percentile CI)
+    # Inform the user to use xxx instead.
+    if(length(ci_colnames) > 2) {
+      cat2(
+        "By default, only one confidence interval is printed."
+      )
+      ci_colnames <- ci_colnames[1:2]
+      cat("\n\n")
+    }
+    
+    ## Total effects
+    cat2("Estimated total effects:\n========================")
+    printEffects(x$Effects$Total_effect, .ci_colnames = ci_colnames, .what = "Total effect")
+    
+    ## Indirect effects
+    cat2("\n\nEstimated indirect effects:\n===========================")
+    printEffects(x$Effects$Indirect_effect, .ci_colnames = ci_colnames, .what = "Indirect effect")
+    
+    ### Variance accounted for -------------------------------------------------
+    cat2("\n\nVariance accounted for (VAF):\n=============================")
+    printEffects(x$Effects$Variance_accounted_for, .ci_colnames = ci_colnames, .what = "Effects")
   }
   
   cat2("\n", rule2(type = 2))