+++ inst/examples/example_predict.R
@@ -51,4 +51,7 @@ res2 <- csem(Anime, # whole data set
 
 # Predict using 10-fold cross-validation with 5 repetitions
 pp2 <- predict(res, .benchmark = "lm")
-pp2}
+pp2
+## There is a plot method available
+plot(pp2)}
+